Configure signing for release build
Here we're creating a keystore with two keys: an upload key and an app signing key. The upload key will be used when signing a bundle for Google Play. The app key will be uploaded to Google Play to be used to sign the final binary. This will let us use the app signing key locally if we need to.