Add internet permission to app
For the website and app to be associated, someone has to visit https://jeran.github.io/.well-known/assetlinks.json.

It's unclear to me if the OS does this, if the password manager does this, or if the app does this.

Just in case it's the app, I'm adding the internet permission.